refactor: reduce per-framework coupling in internal/conventions/ and tests/
Implementing the Dagster framework surfaced two shared-file coupling
points that every framework addition touches, even when the touched
content is framework-exclusive:

- internal/conventions/ grew by 3-4 constants per framework
  (Default<Framework>DBUser/Pass/Name), even though most were only
  ever read by that one framework's own package.
- tests/ mixed framework-specific test functions into shared grab-bag
  files (bootstrap_frameworks_test.go, bootstrap_staged_frameworks_test.go,
  framework_defaults_test.go, framework_snapshot_test.go,
  framework_detection_test.go), so adding a framework meant editing
  files that also contained unrelated frameworks' tests.

Part A: move framework-exclusive constants (DB credential defaults,
a few tool-binary paths) out of internal/conventions/ into the owning
framework's own package, verified single-consumer via grep before each
move. Genuinely shared constants (Magento1/Magento2/MageOS DB defaults,
admin credentials, paths, permissions) stay in conventions. Delete
internal/conventions/frameworks.go (14 FrameworkXxx constants with zero
real consumers) plus two other dead constants (BinWpCli, UserMagento).
Decouple internal/cmd/db_credentials.go's Postgres-engine fallback from
Django's constants by introducing generic DefaultPostgresDBUser/Name.

Part B: reorganize tests/ so a test function exercising exactly one
framework lives in that framework's own <subject>_<framework>_test.go
file; functions that genuinely compare >=2 frameworks are rewritten as
table-driven tests with framework-generic names instead of living in a
shared grab-bag file under a misleadingly specific name. Fold Dagster's
dedicated CA-mount/linked-hosts render test into the existing generic
non-PHP-frameworks loop, closing a duplicate-coverage gap.

Part C: fix Drupal's bootstrap to use the shared
conventions.DefaultAdminUser/Password/Email convention instead of
hardcoded admin/admin/admin@local.test - the one intentional behavior
change here. Existing Drupal sites are unaffected (the password lives
in their database, not in this code path); newly-bootstrapped ones get
the standard Admin12345678$ password instead of admin.

Adding a new framework no longer requires editing internal/conventions/
at all. Every moved test function was verified byte-for-byte identical
to its pre-move body; test-count reconciliation confirms the only
function-count changes are the two deliberate consolidations described
above, with zero accidental coverage loss.

Closes #119
