+++ internal/conventions/binaries.go
@@ -1,9 +1,5 @@
 package conventions
 
 const (
-	BinComposer       = "composer"
-	BinMagento        = "bin/magento"
-	BinArtisan        = "artisan"
-	BinSymfonyConsole = "bin/console"
-	BinWpCli          = "wp"
+	BinComposer = "composer"
 )
