+++ internal/conventions/credentials.go
@@ -20,32 +20,6 @@ const (
 	DefaultMageOSDBPass = "mageos"
 	DefaultMageOSDBName = "mageos"
 
-	DefaultOpenMageDBUser = "openmage"
-	DefaultOpenMageDBPass = "openmage"
-	DefaultOpenMageDBName = "openmage"
-
-	DefaultSymfonyDBUser = "symfony"
-	DefaultSymfonyDBPass = "symfony"
-	DefaultSymfonyDBName = "symfony"
-
-	DefaultLaravelDBUser = "laravel"
-	DefaultLaravelDBPass = "laravel"
-	DefaultLaravelDBName = "laravel"
-
-	DefaultWordPressDBUser = "wordpress"
-	DefaultWordPressDBPass = "wordpress"
-	DefaultWordPressDBName = "wordpress"
-
-	DefaultPrestaShopDBUser      = "prestashop"
-	DefaultPrestaShopDBPass      = "prestashop"
-	DefaultPrestaShopDBName      = "prestashop"
-	DefaultPrestaShopTablePrefix = "ps_"
-
-	DefaultDjangoDBUser = "django"
-	DefaultDjangoDBPass = "django"
-	DefaultDjangoDBName = "django"
-
-	DefaultDagsterDBUser = "dagster"
-	DefaultDagsterDBPass = "dagster"
-	DefaultDagsterDBName = "dagster"
+	DefaultPostgresDBUser = "app"
+	DefaultPostgresDBName = "app"
 )
