--- internal/conventions/frameworks.go
@@ -1,18 +1,0 @@
-package conventions
-
-const (
-	FrameworkMagento1   = "magento1"
-	FrameworkMagento2   = "magento2"
-	FrameworkMageOS     = "mageos"
-	FrameworkLaravel    = "laravel"
-	FrameworkWordPress  = "wordpress"
-	FrameworkSymfony    = "symfony"
-	FrameworkDrupal     = "drupal"
-	FrameworkCakePHP    = "cakephp"
-	FrameworkShopware   = "shopware"
-	FrameworkNextJS     = "nextjs"
-	FrameworkEmdash     = "emdash"
-	FrameworkOpenMage   = "openmage"
-	FrameworkPrestaShop = "prestashop"
-	FrameworkDjango     = "django"
-)
