+++ internal/conventions/users.go
@@ -3,5 +3,4 @@ package conventions
 const (
 	UserWWWData = "www-data"
 	HomeWWWData = "/home/www-data"
-	UserMagento = "magento"
 )
