+++ tests/conventions_django_test.go
@@ -4,20 +4,18 @@ import (
 	"testing"
 
 	"govard/internal/conventions"
+	"govard/internal/frameworks/django"
 )
 
 func TestDjangoConventionsConstants(t *testing.T) {
-	if conventions.FrameworkDjango != "django" {
-		t.Errorf("FrameworkDjango = %q, want %q", conventions.FrameworkDjango, "django")
+	if django.DefaultDBUser != "django" {
+		t.Errorf("DefaultDBUser = %q, want %q", django.DefaultDBUser, "django")
 	}
-	if conventions.DefaultDjangoDBUser != "django" {
-		t.Errorf("DefaultDjangoDBUser = %q, want %q", conventions.DefaultDjangoDBUser, "django")
+	if django.DefaultDBPass != "django" {
+		t.Errorf("DefaultDBPass = %q, want %q", django.DefaultDBPass, "django")
 	}
-	if conventions.DefaultDjangoDBPass != "django" {
-		t.Errorf("DefaultDjangoDBPass = %q, want %q", conventions.DefaultDjangoDBPass, "django")
-	}
-	if conventions.DefaultDjangoDBName != "django" {
-		t.Errorf("DefaultDjangoDBName = %q, want %q", conventions.DefaultDjangoDBName, "django")
+	if django.DefaultDBName != "django" {
+		t.Errorf("DefaultDBName = %q, want %q", django.DefaultDBName, "django")
 	}
 }
