+++ tests/conventions_prestashop_test.go
@@ -4,23 +4,21 @@ import (
 	"testing"
 
 	"govard/internal/conventions"
+	"govard/internal/frameworks/prestashop"
 )
 
 func TestPrestaShopConventionConstants(t *testing.T) {
-	if conventions.FrameworkPrestaShop != "prestashop" {
-		t.Fatalf("expected FrameworkPrestaShop 'prestashop', got %q", conventions.FrameworkPrestaShop)
+	if prestashop.DefaultDBUser != "prestashop" {
+		t.Fatalf("expected DefaultDBUser 'prestashop', got %q", prestashop.DefaultDBUser)
 	}
-	if conventions.DefaultPrestaShopDBUser != "prestashop" {
-		t.Fatalf("expected DefaultPrestaShopDBUser 'prestashop', got %q", conventions.DefaultPrestaShopDBUser)
+	if prestashop.DefaultDBPass != "prestashop" {
+		t.Fatalf("expected DefaultDBPass 'prestashop', got %q", prestashop.DefaultDBPass)
 	}
-	if conventions.DefaultPrestaShopDBPass != "prestashop" {
-		t.Fatalf("expected DefaultPrestaShopDBPass 'prestashop', got %q", conventions.DefaultPrestaShopDBPass)
+	if prestashop.DefaultDBName != "prestashop" {
+		t.Fatalf("expected DefaultDBName 'prestashop', got %q", prestashop.DefaultDBName)
 	}
-	if conventions.DefaultPrestaShopDBName != "prestashop" {
-		t.Fatalf("expected DefaultPrestaShopDBName 'prestashop', got %q", conventions.DefaultPrestaShopDBName)
-	}
-	if conventions.DefaultPrestaShopTablePrefix != "ps_" {
-		t.Fatalf("expected DefaultPrestaShopTablePrefix 'ps_', got %q", conventions.DefaultPrestaShopTablePrefix)
+	if prestashop.DefaultTablePrefix != "ps_" {
+		t.Fatalf("expected DefaultTablePrefix 'ps_', got %q", prestashop.DefaultTablePrefix)
 	}
 	if conventions.PrestaShopParametersFile != "app/config/parameters.php" {
 		t.Fatalf("expected PrestaShopParametersFile 'app/config/parameters.php', got %q", conventions.PrestaShopParametersFile)
