+++ tests/framework_snapshot_test.go
@@ -175,100 +175,18 @@ func TestFrameworkSnapshotBlueprintRendering(t *testing.T) {
 	}
 }
 
-// TestDagsterComposeRendersCAMountAndLinkedProjectHosts is a hermetic,
-// non-snapshot render test covering the "populated" case of Dagster's
-// compose blueprint {{ if .HostGovardRootCAPath }} CA-mount block and
-// {{ range .RuntimeDomainHosts }} extra_hosts block. The golden-file
-// snapshot in TestFrameworkSnapshotBlueprintRendering above always
-// renders with both empty, so it only ever exercised the negative case;
-// this test renders with a real root CA file on disk and a linked
-// project host mapping so both blocks actually populate, and asserts the
-// resulting compose file contains the CA-mount volume line and a
-// linked-project extra_hosts entry (in addition to the always-present
-// static host.docker.internal entry).
-func TestDagsterComposeRendersCAMountAndLinkedProjectHosts(t *testing.T) {
-	tempDir := t.TempDir()
-	fakeHome := filepath.Join(tempDir, "fake-home")
-	if err := os.MkdirAll(fakeHome, 0o755); err != nil {
-		t.Fatalf("failed to create fake home dir: %v", err)
-	}
-	t.Setenv("HOME", fakeHome)
-	t.Setenv("SSH_AUTH_SOCK", "")
-
-	govardHome := filepath.Join(tempDir, "govard-home")
-	t.Setenv("GOVARD_HOME_DIR", govardHome)
-
-	sslDir := filepath.Join(govardHome, "ssl")
-	if err := os.MkdirAll(sslDir, 0o755); err != nil {
-		t.Fatalf("failed to create fake ssl dir: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(sslDir, "root.crt"), []byte("fake-root-ca"), 0o644); err != nil {
-		t.Fatalf("failed to write fake root CA: %v", err)
-	}
-
-	destBlueprintsDir := filepath.Join(tempDir, "blueprints")
-	if err := os.CopyFS(destBlueprintsDir, blueprints.FS); err != nil {
-		t.Fatalf("failed to copy blueprints: %v", err)
-	}
-
-	projectName := "dagster-linked"
-	profileResult, err := engine.ResolveRuntimeProfile("dagster", "")
-	if err != nil {
-		t.Fatalf("ResolveRuntimeProfile failed: %v", err)
-	}
-
-	config := engine.Config{
-		ProjectName:    projectName,
-		Framework:      "dagster",
-		Domain:         projectName + ".test",
-		LinkedProjects: []string{"other-project.test:host-gateway"},
-		Stack: engine.Stack{
-			UserID:  1000,
-			GroupID: 1000,
-			Services: engine.Services{
-				DB: profileResult.Profile.DB,
-			},
-		},
-	}
-	engine.NormalizeConfig(&config, tempDir)
-
-	if err := engine.RenderBlueprint(tempDir, config); err != nil {
-		t.Fatalf("RenderBlueprint failed: %v", err)
-	}
-
-	composePath := engine.ComposeFilePath(tempDir, projectName)
-	composeContent, err := os.ReadFile(composePath)
-	if err != nil {
-		t.Fatalf("failed to read rendered compose file: %v", err)
-	}
-	rendered := string(composeContent)
-
-	if !strings.Contains(rendered, "/usr/local/share/ca-certificates/govard.crt") {
-		t.Errorf("expected rendered Dagster compose to contain the CA-mount volume line, got:\n%s", rendered)
-	}
-	if !strings.Contains(rendered, "other-project.test:host-gateway") {
-		t.Errorf("expected rendered Dagster compose to contain a linked-project extra_hosts entry, got:\n%s", rendered)
-	}
-	if !strings.Contains(rendered, "host.docker.internal:host-gateway") {
-		t.Errorf("expected rendered Dagster compose to still contain the static host.docker.internal extra_hosts entry, got:\n%s", rendered)
-	}
-}
-
 // TestNonPHPFrameworksRenderLinkedProjectExtraHosts is a hermetic,
 // non-snapshot render test covering the "populated" case of the
 // {{ range .RuntimeDomainHosts }} extra_hosts block and the
-// {{ if .HostGovardRootCAPath }} CA-mount block added to Django's,
-// Next.js's, and Emdash's compose blueprints. Before this, none of these
-// frameworks' own web service definitions requested RuntimeDomainHosts or
-// mounted/trusted Govard's root CA - only the shared base.yml's
-// PHP/php-debug services requested RuntimeDomainHosts, and only Dagster's
-// blueprint had the CA-mount mechanism - so linked_projects never
-// resolved to anything but the container's own loopback, and even once it
-// did resolve, outbound HTTPS to a linked project failed TLS verification
-// for any framework outside that set (see the Dagster gap this mirrors,
-// and the render test above covering the same mechanism for Dagster).
+// {{ if .HostGovardRootCAPath }} CA-mount block in Django's, Next.js's,
+// Emdash's, and Dagster's compose blueprints, via a single table-driven
+// loop over all four frameworks. Each of these frameworks' own web service
+// definitions requests RuntimeDomainHosts and mounts/trusts Govard's root
+// CA, so linked_projects resolves to the linked project's runtime domain
+// (not just the container's own loopback) and outbound HTTPS to that
+// linked project passes TLS verification.
 func TestNonPHPFrameworksRenderLinkedProjectExtraHosts(t *testing.T) {
-	for _, framework := range []string{"django", "nextjs", "emdash"} {
+	for _, framework := range []string{"django", "nextjs", "emdash", "dagster"} {
 		framework := framework
 		t.Run(framework, func(t *testing.T) {
 			tempDir := t.TempDir()
